requirements.txt
streamlit
networkx
matplotlib
pandas
google-generativeai
arango
python-dotenv
requests
langchain